pyproject.toml
[project]
name = "ambientvpn-web"
version = "2.1.1"
description = "Ambient VPN Web"
authors = [
    {name = "Denis Tokarev",email = "wf.denis.tokarev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-csp (>=3.8,<4.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "django-stubs (>=5.1.3,<6.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "flake8 (>=7.1.2,<8.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
